requirements.txt
flask
duckdb
apscheduler
requests
pyyaml
pandas
openpyxl
